* android-cmake:
  Android: avoid deprecations aborting the build
  Android: integrate cmake
  Android: disable a few more ARM and MIPS targets
  Android: pass timestamps as a long (int was the only other choice)
  Android: require ANDROID_ARCH to be valid
  Android: correctly fail with the toolchain cannot be found
  Allow ANDROID_NDK_ROOT to be either explicitly set, come from the envrioment, or found on the PATH
  Android: Use $ANDROID_NDK_ROOT, the standard NDK environment varaible

+++ android/RCUtility/app/src/main/jni/tracker_wrapper.cpp
@@ -117,7 +117,7 @@ static void data_callback(void *handle, rc_Timestamp time, rc_Pose pose, rc_Feat
     if (RunExceptionCheck(jniEnv)) return;
 
     // set properties on the SensorFusionData instance
-    CallMethod(jniEnv, dataUpdateClass, dataUpdateObj, "setTimestamp", time);
+    CallMethod(jniEnv, dataUpdateClass, dataUpdateObj, "setTimestamp", (long)time);
     CallMethod(jniEnv, dataUpdateClass, dataUpdateObj, "setPose", pose, 12);
 
     jmethodID methodId = jniEnv->GetMethodID(dataUpdateClass, "addFeaturePoint", "(JFFFFF)V"); // takes a long and 5 floats
